# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ResidentVoice Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /buildings/{buildingId}/residents:
    get:
      operationId: getResidents
      tags: [residents]
      summary: List residents (admin/moderator)
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, verified, rejected, suspended]
      responses:
        "200":
          description: Residents list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResidentUser"
  /buildings/{buildingId}/residents/{userId}/status:
    patch:
      operationId: updateResidentStatus
      tags: [residents]
      summary: Approve, reject, or suspend a resident (admin)
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResidentStatusUpdate"
      responses:
        "200":
          description: Updated resident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResidentUser"

  # ─── Dashboard ───────────────────────────────────────────────────────────────
...
  /buildings/{buildingId}/issues/{issueId}/solutions:
    get:
      operationId: getSolutions
      tags: [solutions]
      summary: List solutions for an issue
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: integer
        - name: issueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Solutions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Solution"
    post:
      operationId: createSolution
      tags: [solutions]
      summary: Suggest a solution
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: integer
        - name: issueId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolutionInput"
      responses:
        "201":
          description: Solution created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Solution"
  /buildings/{buildingId}/issues/{issueId}/solutions/{solutionId}/vote:
    post:
      operationId: voteOnSolution
      tags: [solutions]
      summary: Vote for a solution
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: integer
        - name: issueId
          in: path
          required: true
          schema:
            type: integer
        - name: solutionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vote recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteResult"

  # ─── Resolutions ─────────────────────────────────────────────────────────────
  /buildings/{buildingId}/resolutions:
    get:
      operationId: getResolutions
      tags: [resolutions]
      summary: List resolutions
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Resolutions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resolution"
    post:
      operationId: createResolution
      tags: [resolutions]
      summary: Create a resolution
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolutionInput"
      responses:
        "201":
          description: Resolution created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resolution"
  /buildings/{buildingId}/resolutions/{resolutionId}:
    get:
      operationId: getResolution
      tags: [resolutions]
      summary: Get resolution detail with vote totals
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: integer
        - name: resolutionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resolution detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolutionDetail"
  /buildings/{buildingId}/resolutions/{resolutionId}/vote:
    post:
      operationId: voteOnResolution
      tags: [resolutions]
      summary: Cast a vote on a resolution
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: integer
        - name: resolutionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolutionVoteInput"
      responses:
        "200":
          description: Vote cast
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteResult"
  /buildings/{buildingId}/resolutions/{resolutionId}/status:
    patch:
      operationId: updateResolutionStatus
      tags: [resolutions]
      summary: Open or close voting (admin)
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: integer
        - name: resolutionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolutionStatusUpdate"
      responses:
        "200":
          description: Updated resolution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resolution"

  # ─── Announcements ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ResidentUser:
      type: object
      required: [id, fullName, email, role, verificationStatus, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        verificationStatus:
          type: string
        apartmentNumber:
          type: ["string", "null"]
        floor:
          type: ["integer", "null"]
        residentType:
          type: ["string", "null"]
        createdAt:
          type: string
    ResidentStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [verified, rejected, suspended, pending]
        reason:
          type: string

    # ─── Buildings ─────────────────────────────────────────────────────────────
...
    VoteResult:
      type: object
      required: [count, hasVoted]
      properties:
        count:
          type: integer
        hasVoted:
          type: boolean
        voteType:
          type: ["string", "null"]

    # ─── Solutions ────────────────────────────────────────────────────────────
    Solution:
      type: object
      required: [id, issueId, title, description, voteCount, hasVoted, createdAt]
      properties:
        id:
          type: integer
        issueId:
          type: integer
        title:
          type: string
        description:
          type: string
        estimatedCost:
          type: ["string", "null"]
        pros:
          type: ["string", "null"]
        cons:
          type: ["string", "null"]
        voteCount:
          type: integer
        hasVoted:
          type: boolean
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
    SolutionInput:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
        description:
          type: string
        estimatedCost:
          type: string
        pros:
          type: string
        cons:
          type: string

    # ─── Resolutions ─────────────────────────────────────────────────────────
    Resolution:
      type: object
      required: [id, buildingId, title, description, proposedAction, votingType, eligibleVoters, status, createdAt]
      properties:
        id:
          type: integer
        buildingId:
          type: integer
        title:
          type: string
        description:
          type: string
        proposedAction:
          type: string
        votingType:
          type: string
          enum: [informal, official, petition]
        eligibleVoters:
          type: string
          enum: [owners_only, tenants_only, all_residents, one_per_unit]
        oneVotePerUnit:
          type: boolean
        passPercentage:
          type: integer
        status:
          type: string
          enum: [draft, open, closed, passed, failed]
        relatedIssueId:
          type: ["integer", "null"]
        votingStartAt:
          type: ["string", "null"]
        votingEndAt:
          type: ["string", "null"]
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ResolutionDetail:
      allOf:
        - $ref: "#/components/schemas/Resolution"
        - type: object
          required: [yesCount, noCount, abstainCount, totalVotes, totalEligible, passed]
          properties:
            yesCount:
              type: integer
            noCount:
              type: integer
            abstainCount:
              type: integer
            totalVotes:
              type: integer
            totalEligible:
              type: integer
            passed:
              type: ["boolean", "null"]
            userVote:
              type: ["string", "null"]
    ResolutionInput:
      type: object
      required: [title, description, proposedAction, votingType, eligibleVoters]
      properties:
        title:
          type: string
        description:
          type: string
        proposedAction:
          type: string
        relatedIssueId:
          type: integer
        votingType:
          type: string
          enum: [informal, official, petition]
        eligibleVoters:
          type: string
          enum: [owners_only, tenants_only, all_residents, one_per_unit]
        oneVotePerUnit:
          type: boolean
        passPercentage:
          type: integer
        votingEndAt:
          type: string
    ResolutionVoteInput:
      type: object
      required: [vote]
      properties:
        vote:
          type: string
          enum: [yes, no, abstain]
        comment:
          type: string
    ResolutionStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [open, closed, passed, failed]

    # ─── Announcements ────────────────────────────────────────────────────────
...